* commit 'bdf1f41b7d26e4d07f3e121ed2f973dd8e8f1433':
  disabled oracle in local. enabled oracle for testing in dev
  changed info to debug
  updated readme file
  updated readme file
  multi-db support added functionality to change setting through command line
  multi-db support added functionality to change setting through command line
  multi-db support (disabled cassandra for now)
  multi-db support
  multi-db support
  multi-db support
  multi-db support
  multi-db support
  multi-db support
  multi-db support
  removed import
  cassandra implementation (final).
  cassandra implementation
  cassandra implementation (initial commit).
  cassandra implementation

+++ cakeshop-api/src/main/java/com/jpmorgan/ib/caonpd/cakeshop/db/MetricsBlockListener.java
@@ -130,9 +130,9 @@ public class MetricsBlockListener implements BlockListener, TickListener {
         blockPerMinMeter.mark();
         if (block.getTransactions() != null && block.getTransactions().size() > 0)  {
             if (previousTxCount != null) {
-                long elapsedSec = block.getTimestamp() - previousBlockTime;
+                long elapsedSec = block.getTimestamp().longValue() - previousBlockTime.longValue();
                 currentTxRate = ((double) block.getTransactions().size()) / elapsedSec;
-                pushTxnPerSecRate(block.getTimestamp());
+                pushTxnPerSecRate(block.getTimestamp().longValue());
             }
             txnPerSecMeter.mark(block.getTransactions().size());
             txnPerMinMeter.mark(block.getTransactions().size());
@@ -146,12 +146,12 @@ public class MetricsBlockListener implements BlockListener, TickListener {
 
             if (previousTxCount != null) {
                 currentTxRate = 0.0;
-                pushTxnPerSecRate(block.getTimestamp());
+                pushTxnPerSecRate(block.getTimestamp().longValue());
             }
             previousTxCount = 0;
 
         }
-        previousBlockTime = block.getTimestamp();
+        previousBlockTime = block.getTimestamp().longValue();
     }
 
     @Override
